simplify coco instance datasets
